requirements.txt
flask_socketio
pandas==0.25.3
numpy==1.17.4
imutils==0.5.3
scikit_learn==0.22.1
matplotlib~=3.3.2
opencv-python~=4.4.0.44
Flask~=1.1.2
Pillow~=8.0.1
eventlet~=0.30.0
tensorflow==2.3.1
Keras~=2.4.3
boto3==1.16.47
botocore==1.19.47